Restructure codebase with route groups for marketing/dashboard separation
Major architectural changes:
- Implement Next.js route groups: (marketing) and (dashboard)
- Move landing page to (marketing)/page.tsx
- Move dashboard pages to (dashboard)/dashboard/*
- Create separate layouts for marketing and dashboard sections
- Reorganize components: dashboard/* and marketing/* directories
- Remove unused files and dependencies

Component restructuring:
- Move dashboard sidebar components to components/dashboard/sidebar/
- Move chat features to components/dashboard/features/chat/
- Remove unused shader components (error-fallback, skeleton)
- Remove unused icons directory and preload hook
- Remove unused utilities (color-validation, debounce, error-utils, rate-limit)

Benefits:
- Clear separation between marketing and app code
- Scalable structure for future development
- Cleaner codebase with removed unused files
- Maintains all existing routes and functionality
