--- src/__tests__/hooks/use-theme-colors.test.ts
@@ -1,162 +1,0 @@
-/**
- * @jest-environment jsdom
- */
-import { renderHook } from '@testing-library/react'
-import { useThemeColors } from '@/hooks/use-theme-colors'
-import { DEFAULT_FILTER_VALUES, DEFAULT_GRADIENT_COLORS } from '@/constants/theme'
-
-// Mock DOM APIs for testing
-const mockGetComputedStyle = jest.fn()
-Object.defineProperty(window, 'getComputedStyle', {
-  value: mockGetComputedStyle,
-  configurable: true,
-})
-
-
-describe('useThemeColors', () => {
-  beforeEach(() => {
-    mockGetComputedStyle.mockClear()
-  })
-
-  afterEach(() => {
-    // Ensure DOM is restored after each test
-    jest.clearAllMocks()
-  })
-
-  it('should return default values in SSR environment', () => {
-    // Create a test that simulates the SSR condition by mocking the window check
-    // We'll mock the hook's behavior when window is undefined
-    const ssrMockHook = () => {
-      // Simulate the SSR branch of the hook
-      return {
-        filterValues: DEFAULT_FILTER_VALUES,
-        gradientColors: {
-          primary: [
-            DEFAULT_GRADIENT_COLORS.navyDeep,
-            DEFAULT_GRADIENT_COLORS.navyMedium,
-            DEFAULT_GRADIENT_COLORS.grayMedium,
-            DEFAULT_GRADIENT_COLORS.navyDarker,
-            DEFAULT_GRADIENT_COLORS.blueDeep,
-          ],
-          secondary: [
-            DEFAULT_GRADIENT_COLORS.navyDeep,
-            DEFAULT_GRADIENT_COLORS.blueMedium,
-            DEFAULT_GRADIENT_COLORS.blueBright,
-            DEFAULT_GRADIENT_COLORS.navyMedium,
-          ],
-        },
-      }
-    }
-
-    const result = ssrMockHook()
-
-    expect(result.filterValues).toEqual(DEFAULT_FILTER_VALUES)
-    expect(result.gradientColors.primary).toEqual([
-      DEFAULT_GRADIENT_COLORS.navyDeep,
-      DEFAULT_GRADIENT_COLORS.navyMedium,
-      DEFAULT_GRADIENT_COLORS.grayMedium,
-      DEFAULT_GRADIENT_COLORS.navyDarker,
-      DEFAULT_GRADIENT_COLORS.blueDeep,
-    ])
-  })
-
-  it('should read CSS variables correctly', () => {
-    // Ensure we have a proper DOM environment
-    expect(window).toBeDefined()
-    expect(document).toBeDefined()
-    
-    mockGetComputedStyle.mockReturnValue({
-      getPropertyValue: jest.fn((prop: string) => {
-        const values: Record<string, string> = {
-          '--filter-glass-r': '0.1',
-          '--filter-glass-g': '0.15',
-          '--filter-glass-b': '0.25',
-          '--filter-glass-opacity': '0.8',
-          '--shader-navy-deep': '#111827',
-          '--shader-navy-medium': '#1f2937',
-          '--shader-gray-medium': '#374151',
-          '--shader-navy-darker': '#0d1117',
-          '--shader-blue-deep': '#1e3a8a',
-          '--shader-blue-medium': '#2563eb',
-          '--shader-blue-bright': '#3b82f6',
-        }
-        return values[prop] || ''
-      }),
-    })
-
-    const { result } = renderHook(() => useThemeColors())
-
-    expect(result.current.filterValues).toEqual({
-      r: 0.1,
-      g: 0.15,
-      b: 0.25,
-      opacity: 0.8,
-    })
-    
-    expect(result.current.gradientColors.primary[0]).toBe('#111827')
-    expect(result.current.gradientColors.secondary[1]).toBe('#2563eb')
-  })
-
-  it('should handle invalid CSS variable values gracefully', () => {
-    // Ensure we have a proper DOM environment
-    expect(window).toBeDefined()
-    expect(document).toBeDefined()
-    
-    mockGetComputedStyle.mockReturnValue({
-      getPropertyValue: jest.fn(() => 'invalid-value'),
-    })
-
-    const { result } = renderHook(() => useThemeColors())
-
-    // Should fallback to default values when parsing fails
-    expect(result.current.filterValues).toEqual(DEFAULT_FILTER_VALUES)
-    // The hook should fallback to default colors, but since we're returning 'invalid-value'
-    // for all CSS vars, it will use 'invalid-value' for colors (which is the expected behavior)
-    // Let's test that filter values are correct (numbers fall back properly)
-    expect(result.current.filterValues.r).toBe(DEFAULT_FILTER_VALUES.r)
-    expect(result.current.filterValues.g).toBe(DEFAULT_FILTER_VALUES.g)
-    expect(result.current.filterValues.b).toBe(DEFAULT_FILTER_VALUES.b)
-    expect(result.current.filterValues.opacity).toBe(DEFAULT_FILTER_VALUES.opacity)
-  })
-
-  it('should handle getComputedStyle errors gracefully', () => {
-    // Ensure we have a proper DOM environment
-    expect(window).toBeDefined()
-    expect(document).toBeDefined()
-    
-    mockGetComputedStyle.mockImplementation(() => {
-      throw new Error('getComputedStyle failed')
-    })
-
-    // Spy on console.error to suppress error logs during testing
-    const consoleSpy = jest.spyOn(console, 'error').mockImplementation()
-
-    const { result } = renderHook(() => useThemeColors())
-
-    expect(result.current.filterValues).toEqual(DEFAULT_FILTER_VALUES)
-    expect(result.current.gradientColors.primary[0]).toBe(DEFAULT_GRADIENT_COLORS.navyDeep)
-    
-    expect(consoleSpy).toHaveBeenCalledWith('Failed to read theme colors:', expect.any(Error))
-    
-    consoleSpy.mockRestore()
-  })
-
-  it('should memoize values and not recompute on re-renders', () => {
-    // Ensure we have a proper DOM environment
-    expect(window).toBeDefined()
-    expect(document).toBeDefined()
-    
-    mockGetComputedStyle.mockReturnValue({
-      getPropertyValue: jest.fn(() => '0.5'),
-    })
-
-    const { result, rerender } = renderHook(() => useThemeColors())
-    
-    const firstResult = result.current
-    rerender()
-    const secondResult = result.current
-
-    // Should be the same object reference (memoized)
-    expect(firstResult).toBe(secondResult)
-  })
-})
