+++ src/app/(dashboard)/dashboard/page.tsx
