--- src/components/icons/calendar-check.tsx
@@ -1,110 +1,0 @@
-"use client";
-
-import { motion, useAnimation } from "framer-motion";
-import type { Variants } from "framer-motion";
-import type { HTMLAttributes } from "react";
-import { forwardRef, useCallback, useImperativeHandle, useRef } from "react";
-import { cn } from "@/lib/utils";
-
-export interface CalendarCheckIconHandle {
-  startAnimation: () => void;
-  stopAnimation: () => void;
-}
-
-interface CalendarCheckIconProps extends HTMLAttributes<HTMLDivElement> {
-  size?: number;
-}
-
-const checkVariants: Variants = {
-  normal: {
-    pathLength: 1,
-    opacity: 1,
-    transition: {
-      duration: 0.3,
-    },
-  },
-  animate: {
-    pathLength: [0, 1],
-    opacity: [0, 1],
-    transition: {
-      pathLength: { duration: 0.4, ease: "easeInOut" },
-      opacity: { duration: 0.4, ease: "easeInOut" },
-    },
-  },
-};
-
-const CalendarCheckIcon = forwardRef<
-  CalendarCheckIconHandle,
-  CalendarCheckIconProps
->(({ onMouseEnter, onMouseLeave, className, size = 28, ...props }, ref) => {
-  const controls = useAnimation();
-  const isControlledRef = useRef(false);
-
-  useImperativeHandle(ref, () => {
-    isControlledRef.current = true;
-
-    return {
-      startAnimation: () => controls.start("animate"),
-      stopAnimation: () => controls.start("normal"),
-    };
-  });
-
-  const handleMouseEnter = useCallback(
-    (e: React.MouseEvent<HTMLDivElement>) => {
-      if (!isControlledRef.current) {
-        controls.start("animate");
-      } else {
-        onMouseEnter?.(e);
-      }
-    },
-    [controls, onMouseEnter]
-  );
-
-  const handleMouseLeave = useCallback(
-    (e: React.MouseEvent<HTMLDivElement>) => {
-      if (!isControlledRef.current) {
-        controls.start("normal");
-      } else {
-        onMouseLeave?.(e);
-      }
-    },
-    [controls, onMouseLeave]
-  );
-
-  return (
-    <div
-      className={cn(className)}
-      onMouseEnter={handleMouseEnter}
-      onMouseLeave={handleMouseLeave}
-      {...props}
-    >
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width={size}
-        height={size}
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      >
-        <path d="M8 2v4" />
-        <path d="M16 2v4" />
-        <rect width="18" height="18" x="3" y="4" rx="2" />
-        <path d="M3 10h18" />
-        <motion.path
-          animate={controls}
-          initial="normal"
-          variants={checkVariants}
-          d="m9 16 2 2 4-4"
-          style={{ transformOrigin: "center" }}
-        />
-      </svg>
-    </div>
-  );
-});
-
-CalendarCheckIcon.displayName = "CalendarCheckIcon";
-
-export { CalendarCheckIcon };
