--- src/components/icons/clock-icon.tsx
@@ -1,132 +1,0 @@
-"use client";
-
-import type { Transition, Variants } from "framer-motion";
-import { motion, useAnimation } from "framer-motion";
-import type { HTMLAttributes } from "react";
-import { forwardRef, useCallback, useImperativeHandle, useRef } from "react";
-
-export interface ClockIconHandle {
-  startAnimation: () => void;
-  stopAnimation: () => void;
-}
-
-interface ClockIconProps extends HTMLAttributes<HTMLDivElement> {
-  size?: number;
-}
-
-const handTransition: Transition = {
-  duration: 0.6,
-  ease: [0.4, 0, 0.2, 1],
-};
-
-const handVariants: Variants = {
-  normal: {
-    rotate: 0,
-    originX: "50%",
-    originY: "50%",
-  },
-  animate: {
-    rotate: 360,
-  },
-};
-
-const minuteHandTransition: Transition = {
-  duration: 0.5,
-  ease: "easeInOut",
-};
-
-const minuteHandVariants: Variants = {
-  normal: {
-    rotate: 0,
-    originX: "50%",
-    originY: "50%",
-  },
-  animate: {
-    rotate: 45,
-  },
-};
-
-const ClockIcon = forwardRef<ClockIconHandle, ClockIconProps>(
-  ({ onMouseEnter, onMouseLeave, className, size = 28, ...props }, ref): React.ReactElement => {
-    const controls = useAnimation();
-    const isControlledRef = useRef(false);
-
-    useImperativeHandle(ref, () => {
-      isControlledRef.current = true;
-
-      return {
-        startAnimation: () => controls.start("animate"),
-        stopAnimation: () => controls.start("normal"),
-      };
-    });
-
-    const handleMouseEnter = useCallback(
-      (e: React.MouseEvent<HTMLDivElement>) => {
-        if (!isControlledRef.current) {
-          controls.start("animate");
-        } else {
-          onMouseEnter?.(e);
-        }
-      },
-      [controls, onMouseEnter]
-    );
-
-    const handleMouseLeave = useCallback(
-      (e: React.MouseEvent<HTMLDivElement>) => {
-        if (!isControlledRef.current) {
-          controls.start("normal");
-        } else {
-          onMouseLeave?.(e);
-        }
-      },
-      [controls, onMouseLeave]
-    );
-
-    return (
-      <div
-        className={className}
-        onMouseEnter={handleMouseEnter}
-        onMouseLeave={handleMouseLeave}
-        {...props}
-      >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width={size}
-          height={size}
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        >
-          <circle cx="12" cy="12" r="10" />
-          <motion.line
-            x1="12"
-            y1="12"
-            x2="12"
-            y2="6"
-            variants={handVariants}
-            animate={controls}
-            initial="normal"
-            transition={handTransition}
-          />
-          <motion.line
-            x1="12"
-            y1="12"
-            x2="16"
-            y2="12"
-            variants={minuteHandVariants}
-            animate={controls}
-            initial="normal"
-            transition={minuteHandTransition}
-          />
-        </svg>
-      </div>
-    );
-  }
-);
-
-ClockIcon.displayName = "ClockIcon";
-
-export { ClockIcon };
