--- src/components/icons/logout-icon.tsx
@@ -1,103 +1,0 @@
-"use client";
-
-import type { Variants } from "framer-motion";
-import { motion, useAnimation } from "framer-motion";
-import type { HTMLAttributes } from "react";
-import { forwardRef, useCallback, useImperativeHandle, useRef } from "react";
-
-export interface LogoutIconHandle {
-  startAnimation: () => void;
-  stopAnimation: () => void;
-}
-
-interface LogoutIconProps extends HTMLAttributes<HTMLDivElement> {
-  size?: number;
-}
-
-const pathVariants: Variants = {
-  animate: {
-    x: 2,
-    translateX: [0, -3, 0],
-    transition: {
-      duration: 0.4,
-    },
-  },
-};
-
-const LogoutIcon = forwardRef<LogoutIconHandle, LogoutIconProps>(
-  ({ onMouseEnter, onMouseLeave, className, size = 28, ...props }, ref): React.ReactElement => {
-    const controls = useAnimation();
-    const isControlledRef = useRef(false);
-
-    useImperativeHandle(ref, () => {
-      isControlledRef.current = true;
-
-      return {
-        startAnimation: () => controls.start("animate"),
-        stopAnimation: () => controls.start("normal"),
-      };
-    });
-
-    const handleMouseEnter = useCallback(
-      (e: React.MouseEvent<HTMLDivElement>) => {
-        if (!isControlledRef.current) {
-          controls.start("animate");
-        } else {
-          onMouseEnter?.(e);
-        }
-      },
-      [controls, onMouseEnter]
-    );
-
-    const handleMouseLeave = useCallback(
-      (e: React.MouseEvent<HTMLDivElement>) => {
-        if (!isControlledRef.current) {
-          controls.start("normal");
-        } else {
-          onMouseLeave?.(e);
-        }
-      },
-      [controls, onMouseLeave]
-    );
-
-    return (
-      <div
-        className={className}
-        onMouseEnter={handleMouseEnter}
-        onMouseLeave={handleMouseLeave}
-        {...props}
-      >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width={size}
-          height={size}
-          viewBox="0 0 24 24"
-          fill="none"
-          stroke="currentColor"
-          strokeWidth="2"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        >
-          <path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4" />
-          <motion.polyline
-            points="16 17 21 12 16 7"
-            variants={pathVariants}
-            animate={controls}
-          />
-          <motion.line
-            x1="21"
-            x2="9"
-            y1="12"
-            y2="12"
-            variants={pathVariants}
-            animate={controls}
-          />
-        </svg>
-      </div>
-    );
-  }
-);
-
-LogoutIcon.displayName = "LogoutIcon";
-
-export { LogoutIcon };
