--- src/components/icons/message-square-icon.tsx
@@ -1,105 +1,0 @@
-"use client";
-
-import type { Variants } from "framer-motion";
-import { motion, useAnimation } from "framer-motion";
-import type { HTMLAttributes } from "react";
-import { forwardRef, useCallback, useImperativeHandle, useRef } from "react";
-
-export interface MessageSquareIconHandle {
-  startAnimation: () => void;
-  stopAnimation: () => void;
-}
-
-interface MessageSquareIconProps extends HTMLAttributes<HTMLDivElement> {
-  size?: number;
-}
-
-const iconVariants: Variants = {
-  normal: {
-    scale: 1,
-    rotate: 0,
-  },
-  animate: {
-    scale: 1.05,
-    rotate: [0, -7, 7, 0],
-    transition: {
-      rotate: {
-        duration: 0.5,
-        ease: "easeInOut",
-      },
-      scale: {
-        type: "spring",
-        stiffness: 400,
-        damping: 10,
-      },
-    },
-  },
-};
-
-const MessageSquareIcon = forwardRef<
-  MessageSquareIconHandle,
-  MessageSquareIconProps
->(({ onMouseEnter, onMouseLeave, className, size = 28, ...props }, ref): React.ReactElement => {
-  const controls = useAnimation();
-  const isControlledRef = useRef(false);
-
-  useImperativeHandle(ref, () => {
-    isControlledRef.current = true;
-
-    return {
-      startAnimation: () => controls.start("animate"),
-      stopAnimation: () => controls.start("normal"),
-    };
-  });
-
-  const handleMouseEnter = useCallback(
-    (e: React.MouseEvent<HTMLDivElement>) => {
-      if (!isControlledRef.current) {
-        controls.start("animate");
-      } else {
-        onMouseEnter?.(e);
-      }
-    },
-    [controls, onMouseEnter]
-  );
-
-  const handleMouseLeave = useCallback(
-    (e: React.MouseEvent<HTMLDivElement>) => {
-      if (!isControlledRef.current) {
-        controls.start("normal");
-      } else {
-        onMouseLeave?.(e);
-      }
-    },
-    [controls, onMouseLeave]
-  );
-
-  return (
-    <div
-      className={className}
-      onMouseEnter={handleMouseEnter}
-      onMouseLeave={handleMouseLeave}
-      {...props}
-    >
-      <motion.svg
-        xmlns="http://www.w3.org/2000/svg"
-        width={size}
-        height={size}
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        variants={iconVariants}
-        animate={controls}
-      >
-        <path d="M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z" />
-      </motion.svg>
-    </div>
-  );
-});
-
-MessageSquareIcon.displayName = "MessageSquareIcon";
-
-export { MessageSquareIcon };
